Don't call settimeout() too early.
Calling settimeout too early interferes with IPv6/IPv4 resolution.

"Note: Starting with LuaSocket 3.0, the host name resolution depends on
whether the socket was created by socket.tcp, socket.tcp4 or
socket.tcp6. Addresses from the appropriate family (or both) are tried
in the order returned by the resolver until the first success or until
the last failure. If the timeout was set to zero, only the first address
is tried."